Generate better Struct definitions
Currently, we generate structs as if they are just normal classes that
subclass the `:Struct` superclass, which is not quite correct. We miss
out on generating the getters/setters of the struct members and also end
up not having the right `initialize` method for structs because we don't
(and can't) supply the members to the parent `Struct` class.

The correct way to represent structs is through the:

```ruby
Foo = ::Struct.new(:member1, :member2) do
  def some_method; end
end
```

pattern, which is also the recommended way to create new struct classes.